requirements.txt
flask 
curl
paho-mqtt
